refactor: enhance email verification component
This commit introduces several enhancements to the email verification component:

- Adds the `HasUser` trait to automatically resolve the user.
- Includes a check during component mounting to redirect verified users to the dashboard.
- Removes the manual user authentication check, simplifying the logic.
- Deletes the dedicated feature test `EmailVerificationTest.php` and integrates its functionalities into the `VerifyEmailTest.php` Livewire test, improving test coverage.
- Adds a test case to check if resend verification button works as expected.
