+++ app/Livewire/Auth/VerifyEmail.php
@@ -2,6 +2,7 @@
 
 namespace App\Livewire\Auth;
 
+use App\Concerns\HasUser;
 use App\Models\User;
 use Filament\Forms\Components\Actions;
 use Filament\Forms\Components\Actions\Action;
@@ -9,22 +10,36 @@ use Filament\Forms\Components\Section;
 use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Forms\Contracts\HasForms;
 use Filament\Forms\Form;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Contracts\View\View;
 use Illuminate\Support\Facades\Route;
 use Livewire\Component;
 
 /**
  * @property Form $form
+ * @property User $user
  */
 class VerifyEmail extends Component implements HasForms
 {
+    use HasUser;
     use InteractsWithForms;
+
+    public function mount(): void
+    {
+        if ($this->user->hasVerifiedEmail()) {
+            $this->redirectIntended(default: route('filament.admin.pages.dashboard', absolute: false));
+
+            return;
+        }
+
+        $this->form->fill();
+    }
 
     public function form(Form $form): Form
     {
         return $form
             ->schema([
                 Section::make()
+                    ->key('actions')
                     ->heading(__('filament-panels::pages/auth/email-verification/email-verification-prompt.heading'))
                     ->description(__('Before continuing, could you verify your email address by clicking on the link we just emailed to you? If you didn\'t receive the email, we will gladly send you another.'))
                     ->schema([
@@ -41,20 +56,8 @@ class VerifyEmail extends Component implements HasForms
             ]);
     }
 
-    public function mount(): void
+    public function render(): View
     {
-        $user = Auth::user();
-
-        if (! ($user instanceof User)) {
-            abort(403);
-        }
-
-        if ($user->hasVerifiedEmail()) {
-            $this->redirectIntended(default: route('filament.admin.pages.dashboard', absolute: false));
-
-            return;
-        }
-
-        $this->form->fill();
+        return view('livewire.auth.verify-email');
     }
 }
